refactor(engine): split copy.rs and query/mod.rs to satisfy the line limit
