+++ internal/engine/query/exec.rs
@@ -1,0 +1,225 @@
+//! `exec` command: run a command inside a service container.
+
+use std::io::Write;
+
+use futures_util::StreamExt;
+
+use crate::compose::types::ComposeFile;
+use crate::error::{ComposeError, Result};
+use crate::libpod::types::exec::{
+	ExecCreateConfig, ExecCreateResponse, ExecInspect, ExecStartConfig,
+};
+use crate::libpod::{urlencoded, LogOutput, API_PREFIX};
+
+use super::Engine;
+
+/// Options for [`Engine::exec`], mirroring `docker compose exec` flags.
+#[derive(Default)]
+pub struct ExecOptions {
+	/// Extra environment variables (`KEY=VAL`), `-e/--env`.
+	pub env: Vec<String>,
+	/// Run as this user, `-u/--user`.
+	pub user: Option<String>,
+	/// Working directory inside the container, `-w/--workdir`.
+	pub workdir: Option<String>,
+	/// Run with extended privileges, `--privileged`.
+	pub privileged: bool,
+	/// Detach: start the exec and return without streaming output, `-d/--detach`.
+	pub detach: bool,
+	/// 1-based replica index for a scaled service, `--index` (default: first).
+	pub index: Option<u32>,
+}
+
+/// Build the exec environment list, expanding a bare `KEY` (no `=`) from podup's
+/// own environment the way docker-compose does. A `KEY=VALUE` entry passes
+/// through unchanged; a value-less `KEY` is replaced with `KEY=<host value>` and
+/// dropped entirely when the variable is unset (libpod rejects a bare key with
+/// HTTP 400). Pure so it is unit-tested without a container.
+fn expand_exec_env(env: &[String]) -> Vec<String> {
+	env.iter()
+		.filter_map(|e| {
+			if e.contains('=') {
+				Some(e.clone())
+			} else {
+				std::env::var(e).ok().map(|v| format!("{e}={v}"))
+			}
+		})
+		.collect()
+}
+
+/// True for the in-band stream-teardown line Podman/conmon emits on the exec
+/// stderr channel when an exec launch fails (e.g. a bad `--workdir`/`--user`):
+/// a secondary `read unixpacket ... connection reset by peer` frame that adds
+/// nothing to the real diagnostic. Matching is deliberately narrow so ordinary
+/// program output is never suppressed.
+fn is_exec_teardown_noise(line: &str) -> bool {
+	line.contains("unixpacket") && line.contains("connection reset by peer")
+}
+
+impl Engine {
+	/// Run a command in the first replica of the named service with default
+	/// options. Exits with the command's exit code.
+	pub async fn exec(
+		&self,
+		file: &ComposeFile,
+		service_name: &str,
+		cmd: Vec<String>,
+	) -> Result<()> {
+		self.exec_with_options(file, service_name, cmd, ExecOptions::default())
+			.await
+	}
+
+	/// Run a command in a service container with `docker compose exec`-style
+	/// overrides (env, user, workdir, privileged, detach, replica index).
+	pub async fn exec_with_options(
+		&self,
+		file: &ComposeFile,
+		service_name: &str,
+		cmd: Vec<String>,
+		opts: ExecOptions,
+	) -> Result<()> {
+		// An empty command would be forwarded as an empty `cmd` and surface a raw
+		// podman HTTP 500; reject it up front with a clear message.
+		if cmd.is_empty() {
+			return Err(ComposeError::Unsupported(
+				"exec: a command is required".into(),
+			));
+		}
+		let service = file
+			.services
+			.get(service_name)
+			.ok_or_else(|| ComposeError::ServiceNotFound(service_name.into()))?;
+		// Resolve the target replica through the shared helper so `--index 0` (and
+		// out-of-range indexes) are rejected consistently with `cp`/`port`/etc.,
+		// rather than aliasing index 0 to the first replica.
+		let container_name = self.replica_name_at(service_name, service, opts.index)?;
+
+		let env = expand_exec_env(&opts.env);
+		let exec_cfg = ExecCreateConfig {
+			cmd: Some(cmd),
+			attach_stdout: Some(true),
+			attach_stderr: Some(true),
+			user: opts.user.clone(),
+			working_dir: opts.workdir.clone(),
+			privileged: opts.privileged.then_some(true),
+			env: (!env.is_empty()).then_some(env),
+			..Default::default()
+		};
+		let create_path = format!(
+			"{API_PREFIX}/containers/{}/exec",
+			urlencoded(&container_name),
+		);
+		let resp: ExecCreateResponse = self
+			.client
+			.post_json(&create_path, &exec_cfg)
+			.await
+			.map_err(ComposeError::Podman)?;
+		let exec_id = resp.id;
+
+		// `-d/--detach`: start the exec and return without streaming output or
+		// waiting for the exit code. The server returns immediately, so the
+		// response body is dropped.
+		if opts.detach {
+			let start_cfg = ExecStartConfig {
+				detach: true,
+				tty: false,
+			};
+			let start_path = format!("{API_PREFIX}/exec/{}/start", urlencoded(&exec_id));
+			let _ = self
+				.client
+				.post_json_stream(&start_path, &start_cfg)
+				.await
+				.map_err(ComposeError::Podman)?;
+			return Ok(());
+		}
+
+		let start_cfg = ExecStartConfig {
+			detach: false,
+			tty: false,
+		};
+		let start_path = format!("{API_PREFIX}/exec/{}/start", urlencoded(&exec_id));
+		let start_resp = self
+			.client
+			.post_json_stream(&start_path, &start_cfg)
+			.await
+			.map_err(ComposeError::Podman)?;
+		let mut stream = crate::libpod::parse_multiplexed(start_resp.into_body());
+
+		// Lock stdout once for the whole stream instead of re-acquiring the lock
+		// (and issuing a syscall) per frame; stdout is ours exclusively on this
+		// path. stderr is locked per frame because the tracing subscriber also
+		// writes there: holding its lock across the await loop would starve
+		// concurrent log emissions. Flush after each frame so exec streams
+		// promptly.
+		{
+			let mut out = std::io::stdout().lock();
+			while let Some(msg) = stream.next().await {
+				match msg.map_err(ComposeError::Podman)? {
+					LogOutput::StdOut { message } => {
+						let _ = out.write_all(String::from_utf8_lossy(&message).as_bytes());
+						let _ = out.flush();
+					}
+					LogOutput::StdErr { message } => {
+						let text = String::from_utf8_lossy(&message);
+						// Drop the spurious connection-reset teardown frame an OCI
+						// exec launch-failure emits, so only the real diagnostic shows.
+						if is_exec_teardown_noise(&text) {
+							continue;
+						}
+						let mut err = std::io::stderr().lock();
+						let _ = err.write_all(text.as_bytes());
+						let _ = err.flush();
+					}
+				}
+			}
+		}
+
+		let inspect_path = format!("{API_PREFIX}/exec/{}/json", urlencoded(&exec_id));
+		let inspect: ExecInspect = self
+			.client
+			.get_json(&inspect_path)
+			.await
+			.map_err(ComposeError::Podman)?;
+		if let Some(code) = inspect.exit_code {
+			if code != 0 {
+				return Err(ComposeError::RunExited(code));
+			}
+		}
+
+		Ok(())
+	}
+}
+
+#[cfg(test)]
+mod tests {
+	use super::{expand_exec_env, is_exec_teardown_noise};
+
+	#[test]
+	fn expand_exec_env_passes_through_key_value() {
+		let out = expand_exec_env(&["FOO=bar".to_string(), "BAZ=qux".to_string()]);
+		assert_eq!(out, vec!["FOO=bar".to_string(), "BAZ=qux".to_string()]);
+	}
+
+	#[test]
+	fn expand_exec_env_resolves_bare_key_from_host() {
+		// A bare `KEY` takes its value from podup's own environment; an unset bare
+		// key is dropped (libpod rejects a value-less env entry).
+		std::env::set_var("PODUP_TEST_EXEC_ENV", "from-host");
+		let out = expand_exec_env(&[
+			"PODUP_TEST_EXEC_ENV".to_string(),
+			"PODUP_TEST_EXEC_UNSET_ENV".to_string(),
+		]);
+		std::env::remove_var("PODUP_TEST_EXEC_ENV");
+		assert_eq!(out, vec!["PODUP_TEST_EXEC_ENV=from-host".to_string()]);
+	}
+
+	#[test]
+	fn teardown_noise_matches_only_connection_reset_frame() {
+		assert!(is_exec_teardown_noise(
+			"read unixpacket @->/run/...: read: connection reset by peer"
+		));
+		// Ordinary program output is never suppressed.
+		assert!(!is_exec_teardown_noise("connection reset by peer"));
+		assert!(!is_exec_teardown_noise("hello world"));
+	}
+}
